update func, name and more file
