+++ engine.go
@@ -20,7 +20,6 @@ Package riot is riot engine
 package riot
 
 import (
-	"bytes"
 	"fmt"
 	"log"
 	"os"
@@ -32,8 +31,6 @@ import (
 	"time"
 	// "reflect"
 
-	"encoding/binary"
-	"encoding/gob"
 	"sync/atomic"
 
 	"github.com/go-ego/riot/core"
@@ -745,131 +742,8 @@ func (engine *Engine) Close() {
 	}
 }
 
-// New create a new engine
-func New(dict ...string) *Engine {
-	// func (engine *Engine) New(conf com.Config) *Engine{
-	var (
-		searcher = &Engine{}
-
-		path          = "./riot-index"
-		storageShards = 10
-		numShards     = 10
-
-		segmenterDict string
-	)
-
-	if len(dict) > 0 {
-		segmenterDict = dict[0]
-	}
-
-	searcher.Init(types.EngineOpts{
-		// Using:         using,
-		StorageShards: storageShards,
-		NumShards:     numShards,
-		IndexerOpts: &types.IndexerOpts{
-			IndexType: types.DocIdsIndex,
-		},
-		UseStorage:    true,
-		StorageFolder: path,
-		// StorageEngine: storageEngine,
-		SegmenterDict: segmenterDict,
-		// StopTokenFile: stopTokenFile,
-	})
-
-	// defer searcher.Close()
-	os.MkdirAll(path, 0777)
-
-	// 等待索引刷新完毕
-	// searcher.Flush()
-	// log.Println("recover index number:", searcher.NumDocsIndexed())
-
-	return searcher
-}
-
 // 从文本hash得到要分配到的 shard
 func (engine *Engine) getShard(hash uint32) int {
 	return int(hash - hash/uint32(engine.initOptions.NumShards)*
 		uint32(engine.initOptions.NumShards))
-}
-
-// HasDoc doc is exist return true
-func (engine *Engine) HasDoc(docId uint64) bool {
-	return core.IsDocExist(docId)
-}
-
-// DBHasDoc doc is exist return true
-func (engine *Engine) DBHasDoc(docId uint64) bool {
-	b := make([]byte, 10)
-	length := binary.PutUvarint(b, docId)
-
-	for shard := 0; shard < engine.initOptions.NumShards; shard++ {
-		has, err := engine.dbs[shard].Has(b[0:length])
-		if err != nil {
-			log.Println("engine.dbs[shard].Has(b[0:length]) ", err)
-		}
-
-		if has {
-			return true
-		}
-	}
-
-	return false
-}
-
-// GetDBAllIds get all the DocId from the storage database and return
-// 从数据库遍历所有的 DocId, 并返回
-func (engine *Engine) GetDBAllIds() []uint64 {
-	docsId := make([]uint64, 0)
-	for i := range engine.dbs {
-		engine.dbs[i].ForEach(func(k, v []byte) error {
-			// fmt.Println(k, v)
-			docId, _ := binary.Uvarint(k)
-			docsId = append(docsId, docId)
-			return nil
-		})
-	}
-
-	return docsId
-}
-
-// GetDBAllDocs get the db all docs
-func (engine *Engine) GetDBAllDocs() (
-	docsId []uint64, docData []types.DocIndexData) {
-	for i := range engine.dbs {
-		engine.dbs[i].ForEach(func(key, val []byte) error {
-			// fmt.Println(k, v)
-			docId, _ := binary.Uvarint(key)
-			docsId = append(docsId, docId)
-
-			buf := bytes.NewReader(val)
-			dec := gob.NewDecoder(buf)
-			var data types.DocIndexData
-			err := dec.Decode(&data)
-			if err != nil {
-				log.Println("dec decode ", err)
-			}
-
-			docData = append(docData, data)
-
-			return nil
-		})
-	}
-
-	return docsId, docData
-}
-
-// GetAllDocIds get all the DocId from the storage database and return
-// 从数据库遍历所有的 DocId, 并返回
-func (engine *Engine) GetAllDocIds() []uint64 {
-	return engine.GetDBAllIds()
-}
-
-// Try handler(err)
-func Try(fun func(), handler func(interface{})) {
-	defer func() {
-		if err := recover(); err != nil {
-			handler(err)
-		}
-	}()
-	fun()
 }
