+++ engine_test.go
@@ -42,6 +42,7 @@ func AddDocs(engine *Engine) {
 		Content: "中国十三亿人口",
 		Fields:  ScoringFields{0, 9, 1},
 	})
+
 	engine.Flush()
 }
 
@@ -676,147 +677,6 @@ func TestDocOnlyID(t *testing.T) {
 	engine.Close()
 }
 
-func TestDocRank(t *testing.T) {
-	var engine Engine
-
-	rankOpts := types.RankOpts{
-		ReverseOrder:    true,
-		OutputOffset:    0,
-		MaxOutputs:      1,
-		ScoringCriteria: &RankByTokenProximity{},
-	}
-	engine.Init(types.EngineOpts{
-		Using:           1,
-		IDOnly:          true,
-		SegmenterDict:   "./testdata/test_dict.txt",
-		DefaultRankOpts: &rankOpts,
-		IndexerOpts: &types.IndexerOpts{
-			IndexType: types.LocsIndex,
-		},
-	})
-
-	AddDocs(&engine)
-	engine.RemoveDoc(5)
-	engine.Flush()
-
-	docIds := make(map[uint64]bool)
-	docIds[5] = true
-	docIds[1] = true
-
-	request := types.SearchReq{
-		Text:   "中国人口",
-		DocIds: docIds}
-
-	tokens := engine.Tokens(request)
-	// 建立排序器返回的通信通道
-	rankerReturnChan := make(
-		chan rankerReturnReq, engine.initOptions.NumShards)
-
-	// 生成查找请求
-	lookupRequest := indexerLookupReq{
-		countDocsOnly:    request.CountDocsOnly,
-		tokens:           tokens,
-		labels:           request.Labels,
-		docIds:           request.DocIds,
-		options:          rankOpts,
-		rankerReturnChan: rankerReturnChan,
-		orderless:        request.Orderless,
-		logic:            request.Logic,
-	}
-
-	// 向索引器发送查找请求
-	for shard := 0; shard < engine.initOptions.NumShards; shard++ {
-		engine.indexerLookupChans[shard] <- lookupRequest
-	}
-
-	outputs := engine.Rank(request, rankOpts, tokens, rankerReturnChan)
-
-	if outputs.Docs != nil {
-		outDocs := outputs.Docs.(types.ScoredIDs)
-		utils.Expect(t, "1", len(outDocs))
-	}
-	utils.Expect(t, "2", len(outputs.Tokens))
-	utils.Expect(t, "1", outputs.NumDocs)
-
-	engine.Close()
-}
-
-func TestDocRanks(t *testing.T) {
-	var engine Engine
-
-	rankOpts := types.RankOpts{
-		ReverseOrder:    true,
-		OutputOffset:    0,
-		MaxOutputs:      1,
-		ScoringCriteria: &RankByTokenProximity{},
-	}
-	engine.Init(types.EngineOpts{
-		Using:           1,
-		SegmenterDict:   "./testdata/test_dict.txt",
-		DefaultRankOpts: &rankOpts,
-		IndexerOpts: &types.IndexerOpts{
-			IndexType: types.LocsIndex,
-		},
-	})
-
-	AddDocs(&engine)
-	engine.RemoveDoc(5)
-	engine.Flush()
-
-	docIds := make(map[uint64]bool)
-	docIds[5] = true
-	docIds[1] = true
-
-	request := types.SearchReq{
-		Text:   "中国人口",
-		DocIds: docIds}
-
-	tokens := engine.Tokens(request)
-	// 建立排序器返回的通信通道
-	rankerReturnChan := make(
-		chan rankerReturnReq, engine.initOptions.NumShards)
-
-	// 生成查找请求
-	lookupRequest := indexerLookupReq{
-		countDocsOnly:    request.CountDocsOnly,
-		tokens:           tokens,
-		labels:           request.Labels,
-		docIds:           request.DocIds,
-		options:          rankOpts,
-		rankerReturnChan: rankerReturnChan,
-		orderless:        request.Orderless,
-		logic:            request.Logic,
-	}
-
-	// 向索引器发送查找请求
-	for shard := 0; shard < engine.initOptions.NumShards; shard++ {
-		engine.indexerLookupChans[shard] <- lookupRequest
-	}
-
-	outputs := engine.Ranks(request, rankOpts, tokens, rankerReturnChan)
-
-	if outputs.Docs != nil {
-		outDocs := outputs.Docs.(types.ScoredDocs)
-		utils.Expect(t, "1", len(outDocs))
-	}
-	utils.Expect(t, "2", len(outputs.Tokens))
-	utils.Expect(t, "1", outputs.NumDocs)
-
-	outputs1 := engine.Search(types.SearchReq{
-		Text:    "中国人口",
-		Timeout: 10,
-		DocIds:  docIds})
-
-	if outputs1.Docs != nil {
-		outDocs1 := outputs.Docs.(types.ScoredDocs)
-		utils.Expect(t, "1", len(outDocs1))
-	}
-	utils.Expect(t, "2", len(outputs1.Tokens))
-	utils.Expect(t, "1", outputs1.NumDocs)
-
-	engine.Close()
-}
-
 func TestSearchWithin(t *testing.T) {
 	var engine Engine
 	engine.Init(types.EngineOpts{
@@ -1085,207 +945,4 @@ func TestSearchLogic(t *testing.T) {
 	utils.Expect(t, "[]", outDocs[1].TokenSnippetLocs)
 
 	engine.Close()
-}
-
-func TestDocGetAllDocAndID(t *testing.T) {
-	gob.Register(ScoringFields{})
-
-	var engine Engine
-	engine.Init(types.EngineOpts{
-		Using:         1,
-		NumShards:     2,
-		UseStorage:    true,
-		StorageFolder: "riot.id",
-		IDOnly:        true,
-		SegmenterDict: "./testdata/test_dict.txt",
-		DefaultRankOpts: &types.RankOpts{
-			ReverseOrder:    true,
-			OutputOffset:    0,
-			MaxOutputs:      1,
-			ScoringCriteria: &RankByTokenProximity{},
-		},
-		IndexerOpts: &types.IndexerOpts{
-			IndexType: types.LocsIndex,
-		},
-	})
-
-	AddDocs(&engine)
-	engine.RemoveDoc(5)
-	engine.Flush()
-
-	allIds := engine.GetDBAllIds()
-	fmt.Println("all id", allIds)
-	utils.Expect(t, "4", len(allIds))
-	utils.Expect(t, "[3 4 1 2]", allIds)
-
-	allIds = engine.GetAllDocIds()
-	fmt.Println("all doc id", allIds)
-	utils.Expect(t, "4", len(allIds))
-	utils.Expect(t, "[3 4 1 2]", allIds)
-
-	ids, docs := engine.GetDBAllDocs()
-	fmt.Println("all id and doc", allIds, docs)
-	utils.Expect(t, "4", len(ids))
-	utils.Expect(t, "4", len(docs))
-	utils.Expect(t, "[3 4 1 2]", ids)
-	allDoc := `[{有人口 <nil> [] [] {2 3 1}} {有十三亿人口 <nil> [] [] {2 3 3}} {中国有十三亿人口人口 <nil> [] [] {1 2 3}} {中国人口 <nil> [] [] <nil>}]`
-	utils.Expect(t, allDoc, docs)
-
-	dbhas := engine.DBHasDoc(5)
-	utils.Expect(t, "false", dbhas)
-	// dbhas = engine.DBHasDoc(4)
-	// utils.Expect(t, "true", dbhas)
-
-	docIds := make(map[uint64]bool)
-	docIds[5] = true
-	docIds[1] = true
-	outputs := engine.Search(types.SearchReq{
-		Text:   "中国人口",
-		DocIds: docIds})
-
-	if outputs.Docs != nil {
-		outDocs := outputs.Docs.(types.ScoredIDs)
-		utils.Expect(t, "1", len(outDocs))
-	}
-	utils.Expect(t, "2", len(outputs.Tokens))
-	utils.Expect(t, "1", outputs.NumDocs)
-
-	engine.Close()
-	os.RemoveAll("riot.id")
-}
-
-func TestDocPinYin(t *testing.T) {
-	var engine Engine
-	engine.Init(types.EngineOpts{
-		// Using:         1,
-		UseStorage:    true,
-		StorageFolder: "riot.py",
-		IDOnly:        true,
-		SegmenterDict: "./testdata/test_dict.txt",
-	})
-
-	// AddDocs(&engine)
-	// engine.RemoveDoc(5)
-
-	tokens := engine.PinYin("在路上, in the way")
-
-	fmt.Println("tokens...", tokens)
-	utils.Expect(t, "52", len(tokens))
-
-	var tokenDatas []types.TokenData
-	// tokens := []string{"z", "zl"}
-	for i := 0; i < len(tokens); i++ {
-		tokenData := types.TokenData{Text: tokens[i]}
-		tokenDatas = append(tokenDatas, tokenData)
-	}
-
-	index1 := types.DocIndexData{Tokens: tokenDatas, Fields: "在路上"}
-	index2 := types.DocIndexData{Content: "在路上, in the way",
-		Tokens: tokenDatas}
-
-	engine.IndexDoc(10, index1)
-	engine.IndexDoc(11, index2)
-
-	engine.Flush()
-
-	docIds := make(map[uint64]bool)
-	docIds[5] = true
-	docIds[10] = true
-	docIds[11] = true
-	outputs := engine.Search(types.SearchReq{
-		Text:   "zl",
-		DocIds: docIds,
-	})
-
-	fmt.Println("outputs", outputs.Docs)
-	if outputs.Docs != nil {
-		outDocs := outputs.Docs.(types.ScoredIDs)
-		utils.Expect(t, "2", len(outDocs))
-		// utils.Expect(t, "11", outDocs[0].DocId)
-		// utils.Expect(t, "10", outDocs[1].DocId)
-	}
-	utils.Expect(t, "1", len(outputs.Tokens))
-	utils.Expect(t, "2", outputs.NumDocs)
-
-	engine.Close()
-	os.RemoveAll("riot.py")
-}
-
-func TestForSplitData(t *testing.T) {
-	var engine Engine
-	engine.Init(types.EngineOpts{
-		Using:         4,
-		UseStorage:    true,
-		StorageFolder: "riot.data",
-		IDOnly:        true,
-		SegmenterDict: "./testdata/test_dict.txt",
-	})
-
-	AddDocs(&engine)
-	engine.RemoveDoc(5)
-	engine.Flush()
-
-	tokenDatas := engine.PinYin("在路上, in the way")
-	tokens, num := engine.ForSplitData(tokenDatas, 52)
-	utils.Expect(t, "93", len(tokens))
-	utils.Expect(t, "104", num)
-
-	index1 := types.DocIndexData{Content: "在路上"}
-	engine.IndexDoc(10, index1, true)
-
-	docIds := make(map[uint64]bool)
-	docIds[5] = true
-	docIds[1] = true
-	outputs := engine.Search(types.SearchReq{
-		Text:   "中国人口",
-		DocIds: docIds})
-
-	if outputs.Docs != nil {
-		outDocs := outputs.Docs.(types.ScoredIDs)
-		utils.Expect(t, "0", len(outDocs))
-	}
-	utils.Expect(t, "2", len(outputs.Tokens))
-	utils.Expect(t, "0", outputs.NumDocs)
-
-	engine.Close()
-	os.RemoveAll("riot.data")
-}
-
-func TestDocCounters(t *testing.T) {
-	var engine Engine
-	engine.Init(types.EngineOpts{
-		Using:         1,
-		UseStorage:    true,
-		StorageFolder: "riot.doc",
-		IDOnly:        true,
-		SegmenterDict: "./testdata/test_dict.txt",
-	})
-
-	AddDocs(&engine)
-	engine.RemoveDoc(5)
-	engine.Flush()
-
-	numAdd := engine.NumTokenIndexAdded()
-	utils.Expect(t, "14", numAdd)
-	numInx := engine.NumDocsIndexed()
-	utils.Expect(t, "5", numInx)
-	numRm := engine.NumDocsRemoved()
-	utils.Expect(t, "8", numRm)
-
-	docIds := make(map[uint64]bool)
-	docIds[5] = true
-	docIds[1] = true
-	outputs := engine.Search(types.SearchReq{
-		Text:   "中国人口",
-		DocIds: docIds})
-
-	if outputs.Docs != nil {
-		outDocs := outputs.Docs.(types.ScoredIDs)
-		utils.Expect(t, "1", len(outDocs))
-	}
-	utils.Expect(t, "2", len(outputs.Tokens))
-	utils.Expect(t, "1", outputs.NumDocs)
-
-	engine.Close()
-	os.RemoveAll("riot.doc")
 }
